moved first and second fragment into UI file as it is part of the UI.
